requirements.txt
more_itertools
numpy
torch
tqdm
parameterized
pytest
pyyaml
retry
tensorboardX
wandb
git+https://github.com/Hiroshiba/pytorch-trainer@1.3.1
git+https://github.com/Hiroshiba/acoustic_feature_extractor@7cab860d9623c34a6ce818ff87c0c14a9ed60231
